enemy hitbox, player health, death code, attempt at path for enemy, fixed attack anim on player
added code so that the player can lose health and die, it isn't running
because i couldn't get the enemy to make it lose health properly, but
there is code to make him lose health on the event tick so you can see
it. Added a broken enemy that is meant to walk on a path, doesn't work.
Added hit box to original enemy to run the lose hp function on the
player. Fixed issue where player could spin around during attack
animation